secboot,cmd/snap-bootstrap: move connections, unlock logic to secboot
Move TPM-dependant unlock logic to secboot so that we call a higher level
function to unlock an encrypted device. This gives us more flexibility
to deal with unlocking with TrustZone-based solutions that can have
arbitrary APIs.
