+++ secboot/secboot.go
@@ -23,3 +23,22 @@ package secboot
 // the github.com/snapcore/secboot repository. That will ensure
 // it can be build as part of the debian build without secboot.
 // Debian does run "go list" without any support for passing -tags.
+
+import (
+	"path/filepath"
+
+	"github.com/snapcore/snapd/osutil"
+)
+
+var (
+	// for mocking by tests
+	devDiskByLabelDir = "/dev/disk/by-label"
+)
+
+func isDeviceEncrypted(name string) (ok bool, encdev string) {
+	encdev = filepath.Join(devDiskByLabelDir, name+"-enc")
+	if osutil.FileExists(encdev) {
+		return true, encdev
+	}
+	return false, ""
+}
