+++ secboot/secboot_tpm.go
@@ -24,10 +24,13 @@ import (
 	"errors"
 	"fmt"
 	"os"
+	"path/filepath"
 
 	sb "github.com/snapcore/secboot"
+	"golang.org/x/xerrors"
 
 	"github.com/snapcore/snapd/asserts"
+	"github.com/snapcore/snapd/boot"
 	"github.com/snapcore/snapd/bootloader/efi"
 	"github.com/snapcore/snapd/logger"
 )
@@ -36,6 +39,7 @@ var (
 	sbConnectToDefaultTPM            = sb.ConnectToDefaultTPM
 	sbMeasureSnapSystemEpochToTPM    = sb.MeasureSnapSystemEpochToTPM
 	sbMeasureSnapModelToTPM          = sb.MeasureSnapModelToTPM
+	sbLockAccessToSealedKeys         = sb.LockAccessToSealedKeys
 	sbActivateVolumeWithTPMSealedKey = sb.ActivateVolumeWithTPMSealedKey
 )
 
@@ -87,6 +91,7 @@ func NewTPMFromConnection(tpm *sb.TPMConnection) *TPM {
 	return &TPM{tpm: tpm}
 }
 
+/*
 func SecureConnect(ekcfile string) (*TPM, error) {
 	ekCertReader, err := os.Open(ekcfile)
 	if err != nil {
@@ -100,8 +105,11 @@ func SecureConnect(ekcfile string) (*TPM, error) {
 	}
 	return &TPM{tpm: tpm}, nil
 }
+*/
 
-func InsecureConnect() (*TPM, error) {
+var insecureConnect = insecureConnectImpl
+
+func insecureConnectImpl() (*TPM, error) {
 	tpm, err := sb.ConnectToDefaultTPM()
 	if err != nil {
 		return nil, err
@@ -109,35 +117,16 @@ func InsecureConnect() (*TPM, error) {
 	return &TPM{tpm: tpm}, nil
 }
 
+func MockInsecureConnect(f func() (*TPM, error)) (restore func()) {
+	old := insecureConnect
+	insecureConnect = f
+	return func() {
+		insecureConnect = old
+	}
+}
+
 func (t *TPM) Disconnect() error {
 	return t.tpm.Close()
-}
-
-func LockAccessToSealedKeys(t *TPM) error {
-	return sb.LockAccessToSealedKeys(t.tpm)
-}
-
-// UnlockEncryptedPartition unseals the keyfile and opens an encrypted device.
-func UnlockEncryptedPartition(t *TPM, name, device, keyfile, pinfile string, lock bool) error {
-	options := sb.ActivateWithTPMSealedKeyOptions{
-		PINTries:            1,
-		RecoveryKeyTries:    3,
-		LockSealedKeyAccess: lock,
-	}
-
-	// XXX: pinfile is currently not used
-	activated, err := sbActivateVolumeWithTPMSealedKey(t.tpm, name, device, keyfile, nil, &options)
-	if !activated {
-		// ActivateVolumeWithTPMSealedKey should always return an error if activated == false
-		return fmt.Errorf("cannot activate encrypted device %q: %v", device, err)
-	}
-	if err != nil {
-		logger.Noticef("successfully activated encrypted device %q using a fallback activation method", device)
-	} else {
-		logger.Noticef("successfully activated encrypted device %q with TPM", device)
-	}
-
-	return nil
 }
 
 func MeasureEpoch(t *TPM) error {
@@ -151,5 +140,106 @@ func MeasureModel(t *TPM, model *asserts.Model) error {
 	if err := sbMeasureSnapModelToTPM(t.tpm, tpmPCR, model); err != nil {
 		return fmt.Errorf("cannot measure snap system model: %v", err)
 	}
+	return nil
+}
+
+func MeasureWhenPossible(whatHow func(t *TPM) error) error {
+	// the model is ready, we're good to try measuring it now
+	t, err := insecureConnect()
+	if err != nil {
+		var perr *os.PathError
+		// XXX: xerrors.Is() does not work with PathErrors?
+		if xerrors.As(err, &perr) {
+			// no TPM
+			return nil
+		}
+		return fmt.Errorf("cannot open TPM connection: %v", err)
+	}
+	defer t.Disconnect()
+
+	return whatHow(t)
+}
+
+// UnlockIfEncrypted verifies whether an encrypted volume with the specified
+// name exists and unlocks it. With lockKeysOnFinish set, access to the sealed
+// keys will be locked when this function completes. The path to the unencrypted
+// device node is returned.
+func UnlockIfEncrypted(name string, lockKeysOnFinish bool) (string, error) {
+	device := filepath.Join(devDiskByLabelDir, name)
+
+	// TODO:UC20: use sb.SecureConnectToDefaultTPM() if we decide there's benefit in doing that or
+	//            we have a hard requirement for a valid EK cert chain for every boot (ie, panic
+	//            if there isn't one). But we can't do that as long as we need to download
+	//            intermediate certs from the manufacturer.
+	tpm, tpmErr := sbConnectToDefaultTPM()
+	if tpmErr != nil {
+		logger.Noticef("cannot open TPM connection: %v", tpmErr)
+	} else {
+		defer tpm.Close()
+	}
+
+	var lockErr error
+	err := func() error {
+		defer func() {
+			// TODO:UC20: we might want some better error handling here - eg, if tpmErr is a
+			//            *os.PathError returned from go-tpm2 then this is an indicator that there
+			//            is no TPM device. But other errors probably shouldn't be ignored.
+			if lockKeysOnFinish && tpmErr == nil {
+				// Lock access to the sealed keys. This should be called whenever there
+				// is a TPM device detected, regardless of whether secure boot is enabled
+				// or there is an encrypted volume to unlock. Note that snap-bootstrap can
+				// be called several times during initialization, and if there are multiple
+				// volumes to unlock we should lock access to the sealed keys only after
+				// the last encrypted volume is unlocked, in which case lockKeysOnFinish
+				// should be set to true.
+				lockErr = sbLockAccessToSealedKeys(tpm)
+			}
+		}()
+
+		ok, encdev := isDeviceEncrypted(name)
+		if !ok {
+			return nil
+		}
+
+		if tpmErr != nil {
+			return fmt.Errorf("cannot unlock encrypted device %q: %v", name, tpmErr)
+		}
+		// TODO:UC20: snap-bootstrap should validate that <name>-enc is what
+		//            we expect (and not e.g. an external disk), and also that
+		//            <name> is from <name>-enc and not an unencrypted partition
+		//            with the same name (LP #1863886)
+		sealedKeyPath := filepath.Join(boot.InitramfsEncryptionKeyDir, name+".sealed-key")
+		return unlockEncryptedPartition(tpm, name, encdev, sealedKeyPath, "", lockKeysOnFinish)
+	}()
+	if err != nil {
+		return "", err
+	}
+	if lockErr != nil {
+		return "", fmt.Errorf("cannot lock access to sealed keys: %v", lockErr)
+	}
+
+	return device, nil
+}
+
+// UnlockEncryptedPartition unseals the keyfile and opens an encrypted device.
+func unlockEncryptedPartition(tpm *sb.TPMConnection, name, device, keyfile, pinfile string, lock bool) error {
+	options := sb.ActivateWithTPMSealedKeyOptions{
+		PINTries:            1,
+		RecoveryKeyTries:    3,
+		LockSealedKeyAccess: lock,
+	}
+
+	// XXX: pinfile is currently not used
+	activated, err := sbActivateVolumeWithTPMSealedKey(tpm, name, device, keyfile, nil, &options)
+	if !activated {
+		// ActivateVolumeWithTPMSealedKey should always return an error if activated == false
+		return fmt.Errorf("cannot activate encrypted device %q: %v", device, err)
+	}
+	if err != nil {
+		logger.Noticef("successfully activated encrypted device %q using a fallback activation method", device)
+	} else {
+		logger.Noticef("successfully activated encrypted device %q with TPM", device)
+	}
+
 	return nil
 }
